shells/nushell: update to 0.68.1
This is the 0.68.1 release of Nushell. It is a bugfix release of 0.68.
This addresses a bug with blocks and source-env.

For convenience, we are providing full builds for Windows, Linux, and macOS.
These are the "all extra features" builds, so be sure you have the requirements
to enable all capabilities:
https://www.nushell.sh/book/installation.html#dependencies

With this release, we have also statically linked OpenSSL into the binary to
make it easier to use across a larger range of platforms. If you encounter any
issue with this, please let us know.