gelemental: update to 2.0.1.
Fixes build.

Version 2.0.1 [stable]
======================

This is a maintenance release of gElemental which is C++17 compatible and
forces button colours for several modern themes.

Version 2.0.0 [stable]
----------------------

This release of gElemental reflects the transfer to a new maintainer.
It includes several bugfixes from the Debichem Team, some minor interface
improvements, and the addition of new element names.