pyversion.mk: Define PYPKGPREFIX in error case
The specification for pyversion.mk says that it defines PYPKGPREFIX
and PYVERSSUFFIX.  Various buildlink files rely on these being
defined, as they are tested without guarding them for being empty.

When there is no valid python version, _PYTHON_VERSION was set to
"none", and PKG_FAIL_REASON defined, so the user gets a reasonable
error.  However, if a buildlink3 uses an unguarded PYPKGPREFIX, a
syntax error results, and the PKG_FAIL_REASON is not display.

This commit defines the two variables to "none" in the case of no
valid version, mirroring the treatment of _PYTHON_VERSION and
correcting a failure to follow the specification.  In this case the
build is going to fail one way or another, but it's vastly better to
have a useful error message.

(Tested earlier, but deferred during freeze.)