openssl: downgrade to 1.1.1q
The tarball was retracted due to a regression, to quote:

We have received a report of a significant regression in the latest
3.0.6 and 1.1.1r versions. The regression is not thought to have
security consequences. While the regression is further investigated we
have taken the decision to withdraw the 3.0.6 and 1.1.1r versions and
instead recommend that users remain on the previous 3.0.5 and 1.1.1q
versions for now.

We will issue a new plan for the release of 3.0.7 and 1.1.1s soon.

From https://mta.openssl.org/pipermail/openssl-announce/2022-October/000237.html