py-hypothesis: updated to 6.56.3
6.56.3 - 2022-10-17

This patch teaches text() to rewrite a few more filter predicates. You’re unlikely to notice any change.

6.56.2 - 2022-10-10

This patch updates our vendored list of top-level domains, which is used by the provisional domains() strategy, and fixes some incorrect examples in the docs for mutually_broadcastable_shapes().

6.56.1 - 2022-10-05

This patch improves the error message when Hypothesis detects “flush to zero” mode for floating-point: we now report which package(s) enabled this, which can make debugging much easier.

6.56.0 - 2022-10-02

This release defines __bool__() on SearchStrategy. It always returns True, like before, but also emits a warning to help with cases where you intended to draw a value.

6.55.0 - 2022-09-29

In preparation for future versions of the Array API standard, make_strategies_namespace() now accepts an optional api_version argument, which determines the version conformed to by the returned strategies namespace. If None, the version of the passed array module xp is inferred.

This release also introduces xps.real_dtypes(). This is currently equivalent to the existing xps.numeric_dtypes() strategy, but exists because the latter is expected to include complex numbers in the next version of the standard.