py37-reduce: Drop 37 from PYTHON_VERSIONS_ACCEPTED
Drop 37 from PYTHON_VERSIONS_ACCEPTED for all but
py-backports.cached-property (which exists only for 37, and I'm
guessing it will build fine and not cause trouble).

This follows the new plan of "37 is all but removed, with the
lang/python37 package remaining, but py37-foo are not built by
default" compromise to avoid problems from the current state of
massive py37-foo build failures while stopping short of outright
removal.